gogensig/refactor:size move to type.go
